Updated Program
Fixed numerous minor bugs.

Total IBI counts for original and raw file are not correct in case processing file.

Total amount of time (as a percent of the total analyzed file's time) subjected to Bayesian GP is now accurate in case processing summary file as well.

Finally minor bug with the divide function was fixed.